Added new KeyNotFoundException
This new exception extends SnappyDbException so the change should be backward compatible, but it now allows to differentiate between an actual error and a key not found. It may not be the best solution to have an exception for something that is not exceptional but it's the easiest way to support primitive types returned from the get* methods, otherwise a switch to an Optional type could also have been possible.